// src/app/models/UsTranslations.ts
export interface UsTranslations {
  bannerTitle: string;
  bannerDescription: string;
  email: string;
  name: string;
  message: string;
  phone: string;
  contactUs: string;
  whoWeAre: string;
  whoWeAreDescription1: string;
  whoWeAreDescription2: string;
  value1: string;
  value1Description: string;
  value2: string;
  value2Description: string;
  value3: string;
  value3Description: string;
  value4: string;
  value4Description: string;
  serviceTitle: string;
  design: string;
  remodeling: string;
  achitecture: string;
  seeMore: string;
  contactDescription: string;
  contactInvitation: string;
  thanksForContacting: string;
  contactSoon: string;
}